refactor(LottoWinningNumbers, LottoMatcher): LottoWinningNumbers가 티켓과의 비교 기능을 담당하도록 변경
LottoMatcher라는 별도의 유틸성 클래스를 만들기 보다는,
LottoWinningNumbers 객체에 로또 티켓 객체들을 전달하여 비교 결과를 요청하는 것으로 변경
